Infer nfolds directly in printLassoRMSE function
Currently, the nfolds in printLassoRMSE are read from the global variable nfolds. This means if the notebook user changes the value for cv directly in the lassocv function rather than changing the global variable, it won't update to the value the user set and when the user runs printLassoRMSE, it will only print the first 3 values. By adding a line to the function definition to infer the value for nfolds directly from the model object, this problem is avoided and makes the notebook a bit more user friendly for beginners. 

The code:

cvlasso.mse_path_.shape[-1] 

Returns the length of the last dimension of the array. So for a 1d array (when alpha is set to a single value), it will return the length of the 1d array which is equal to the number of folds. For a 2d array (when alpha is set to an array of regularization parameters to try), it will return the length of the first column of the 2d array which is also equal to the number of folds. 

mse_path_ is an attribute of the lassocv object of shape (nalphas, nfolds)

See documentation here: https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LassoCV.html for more detail.